Fix file import errors, Add auto-mounting, remove some trailing whitespace
I'm not sure as to why this is, but you were importing with ./ in the beginning and that broke the lame garry file system. I just got rid of that and this currently works. Also got rid of a decent portion of trailing whitespace and added auto mounting by searching the materials folder. I really like this project so I just wanna make sure it works lol.

Also cause I wanna get it to be an editor for Starfall / E2 / Any other chip instead of the garbo ace editor.